# backend/openapi.yaml
openapi: 3.1.0
info:
  title: WhatsApp AI Agent — Multi-Tenant SaaS API
  version: 1.0.0
  description: |
    ## Overview
    Multi-tenant restaurant ordering platform powered by an AI agent over WhatsApp.
    Each **Business** (tenant) connects a WhatsApp phone number, and the AI agent
    handles customer orders, menu inquiries, complaints, and status checks automatically.

    **Two API surfaces:**
    | Surface | Auth | Scope |
    |---------|------|-------|
    | **Tenant API** | `x-api-key` or Bearer JWT | Scoped to one business |
    | **Platform API** | Bearer JWT only (`businessId: null`) | Cross-tenant super-admin |

    ## Webhook Flow
    1. Meta sends a WhatsApp message to `POST /api/webhook`
    2. Server identifies the Business via `phone_number_id`
    3. AI agent processes the message (menu query, order submission, etc.)
    4. Reply is sent back as the response body

    ## Real-time Events (Socket.io)
    - **URL:** `http://localhost:3001`
    - **Rooms:** `business:{businessId}`
    - **Events:**
      | Event | Payload | Description |
      |-------|---------|-------------|
      | `new-message` | `{ customerId, customerPhone, message, reply }` | AI replied to a customer |
      | `new-order` | `{ orderId, referenceId, customerId, items, totalPrice }` | New order placed via WhatsApp |
      | `order-updated` | `{ orderId, status, updatedAt }` | Order status changed |
...

    ## Demo Credentials
    | Role | API Key / Endpoint | Email | Password |
...
servers:
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Business API key (e.g. `demo-api-key-123`)
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: User access token from `/api/auth/login`
    PlatformBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Platform user JWT from `/api/auth/platform-login`.
        Token contains `businessId: null`.
  schemas:
...
    MetaWebhookPayload:
      type: object
      required: [object, entry]
      properties:
        object:
          type: string
          example: whatsapp_business_account
        entry:
          type: array
          items:
            $ref: '#/components/schemas/WebhookEntry'
    WebhookEntry:
      type: object
      properties:
        id:
          type: string
        changes:
          type: array
          items:
            $ref: '#/components/schemas/WebhookChange'
    WebhookChange:
      type: object
      properties:
        value:
          $ref: '#/components/schemas/WebhookValue'
        field:
          type: string
          example: messages
    WebhookValue:
      type: object
      properties:
        messaging_product:
          type: string
          example: whatsapp
        metadata:
          type: object
          properties:
            display_phone_number:
              type: string
            phone_number_id:
              type: string
        contacts:
          type: array
          items:
            type: object
            properties:
              profile:
                type: object
                properties:
                  name:
                    type: string
              wa_id:
                type: string
        messages:
          type: array
          items:
            $ref: '#/components/schemas/WebhookMessage'
    WebhookMessage:
      type: object
      properties:
        from:
          type: string
        id:
          type: string
        timestamp:
          type: string
        type:
          type: string
          example: text
        text:
          type: object
          properties:
            body:
              type: string

    # ─── IAM ───
...
paths:
...
  /webhook:
    get:
      tags: [Webhook]
      summary: Meta webhook verification
      description: |-
        Meta sends a GET request to verify the webhook endpoint.
        Responds with the `hub.challenge` value if `hub.verify_token` matches.
      security: []
      parameters:
        - name: hub.mode
          in: query
          required: true
          schema:
            type: string
            example: subscribe
        - name: hub.verify_token
          in: query
          required: true
          schema:
            type: string
        - name: hub.challenge
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Verification successful
          content:
            text/plain:
              schema:
                type: string
        '403':
          description: Invalid verify token
    post:
      tags: [Webhook]
      summary: Receive WhatsApp message
      description: |-
        Receives incoming WhatsApp messages from Meta's Cloud API.
        The AI agent processes the message and returns a reply.
        Business is identified by `phone_number_id` in the payload.
      security: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MetaWebhookPayload'
      responses:
        '200':
          description: AI agent reply
          content:
            application/json:
              schema:
                type: object
                properties:
                  reply:
                    type: string
                    description: The AI agent's text response
        '404':
          description: Business not found for the given phone_number_id

  # ─── AUTH ────────────────────────────────────────────────────────────────────
...